+++ superdesk/media/renditions.py
@@ -16,9 +16,12 @@ from copy import deepcopy
 from flask import current_app as app
 from .media_operations import process_file_from_stream
 from .media_operations import crop_image
+from .media_operations import download_file_from_url
+from .media_operations import process_file
 from .image import fix_orientation
 from eve.utils import config
 from superdesk import get_resource_service
+from superdesk.filemeta import set_filemeta
 
 
 logger = logging.getLogger(__name__)
@@ -38,7 +41,7 @@ def generate_renditions(original, media_id, inserted, file_type, content_type,
     :param bool insert_metadata: boolean to inserted metadata or not. For AWS storage it is false.
     :return: dict of renditions
     """
-    rend = {'href': url_for_media(media_id, content_type), 'media': media_id, 'mimetype': content_type}
+    rend = {'href': app.media.url_for_media(media_id, content_type), 'media': media_id, 'mimetype': content_type}
     renditions = {'original': rend}
 
     if file_type != 'image':
@@ -238,3 +241,45 @@ def get_renditions_spec(without_internal_renditions=False, no_custom_crops=False
                 # complete list of wanted renditions
                 rendition_spec[crop['name']] = crop
     return rendition_spec
+
+
+def update_renditions(item, href, old_item):
+    """Update renditions for an item.
+
+    If the old_item has renditions uploaded in to media then the old rendition details are
+    assigned to the item, this avoids repeatedly downloading the same image and leaving the media entries orphaned.
+    If there is no old_item the original is downloaded and renditions are
+    generated.
+    :param item: parsed item from source
+    :param href: reference to original
+    :param old_item: the item that we have already ingested, if it exists
+    :return: item with renditions
+    """
+    inserted = []
+    try:
+        # If there is an existing set of renditions we keep those
+        if old_item:
+            media = old_item.get('renditions', {}).get('original', {}).get('media', {})
+            if media:
+                item['renditions'] = old_item['renditions']
+                item['mimetype'] = old_item.get('mimetype')
+                item['filemeta'] = old_item.get('filemeta')
+                item['filemeta_json'] = old_item.get('filemeta_json')
+                return
+
+        content, filename, content_type = download_file_from_url(href)
+        file_type, ext = content_type.split('/')
+        metadata = process_file(content, file_type)
+        file_guid = app.media.put(content, filename, content_type, metadata)
+        inserted.append(file_guid)
+        rendition_spec = get_renditions_spec()
+        renditions = generate_renditions(content, file_guid, inserted, file_type,
+                                         content_type, rendition_spec, app.media.url_for_media)
+        item['renditions'] = renditions
+        item['mimetype'] = content_type
+        set_filemeta(item, metadata)
+    except Exception as e:
+        logger.exception(e)
+        for file_id in inserted:
+            app.media.delete(file_id)
+        raise
